refactor(idl): centralise Candid IDL factories under src/declarations/ (#350)
Eliminates the three-place update rule documented in CLAUDE.md: inline
idlFactory functions previously lived in each of the 16 service files and
had to be kept in sync manually.  Each canister now has a single source of
truth at frontend/src/declarations/<canister>/index.ts; service files
import and re-export idlFactory from there.

- Create frontend/src/declarations/<canister>/index.ts for all 16 canisters
  (auth, job, property, contractor, quote, payment, photo, report,
  maintenance, sensor, listing, recurring, monitoring, bills, ai_proxy,
  icp_ledger)
- Strip inline idlFactory from every service file; replace with import +
  re-export so existing contract tests continue to import from service paths
- Add scripts/generate-declarations.sh — one-command regeneration via
  `dfx generate` once dfx is installed
- Update CLAUDE.md: IDL maintenance rule now points to declarations/; drop
  the "three-place update" wording

All 3094 unit tests pass.

Closes #350
