+++ frontend/src/services/maintenance.ts
@@ -7,43 +7,10 @@
 import { Actor } from "@icp-sdk/core/agent";
 import { getAgent } from "./actor";
 import type { Job } from "./job";
+import { idlFactory } from "@/declarations/maintenance";
+export { idlFactory };
 
 const MAINTENANCE_CANISTER_ID = (process.env as any).MAINTENANCE_CANISTER_ID || "";
-
-// ─── IDL (schedule methods only) ─────────────────────────────────────────────
-
-export const idlFactory = ({ IDL }: any) => {
-  const ScheduleEntry = IDL.Record({
-    id:                 IDL.Text,
-    propertyId:         IDL.Text,
-    systemName:         IDL.Text,
-    taskDescription:    IDL.Text,
-    plannedYear:        IDL.Nat,
-    plannedMonth:       IDL.Opt(IDL.Nat),
-    estimatedCostCents: IDL.Opt(IDL.Nat),
-    isCompleted:        IDL.Bool,
-    createdBy:          IDL.Principal,
-    createdAt:          IDL.Int,
-  });
-  const Error = IDL.Variant({
-    NotFound:     IDL.Null,
-    NotAuthorized: IDL.Null,
-    InvalidInput: IDL.Text,
-  });
-  return IDL.Service({
-    createScheduleEntry: IDL.Func(
-      [IDL.Text, IDL.Text, IDL.Text, IDL.Nat, IDL.Opt(IDL.Nat), IDL.Opt(IDL.Nat)],
-      [IDL.Variant({ ok: ScheduleEntry, err: Error })],
-      []
-    ),
-    getScheduleByProperty: IDL.Func([IDL.Text], [IDL.Vec(ScheduleEntry)], ["query"]),
-    markCompleted: IDL.Func(
-      [IDL.Text],
-      [IDL.Variant({ ok: ScheduleEntry, err: Error })],
-      []
-    ),
-  });
-};
 
 function fromEntry(raw: any): ScheduleEntry {
   return {
